commented out code on ListView, created ListForm
